projects/final/Kucia_Grabek_Trochimiak/requirements.txt
streamlit
openai
audio_recorder_streamlit
streamlit-float

PyMuPDF
nltk
pyyaml
tqdm

sentence_transformers
chromadb
spacy
cupy-cuda12x
torch

flask
